refactor : PlaylistAdapter 위치 변경
