+++ android/core/ui/src/main/res/values/strings.xml
@@ -34,7 +34,7 @@
     <string name="description">설명</string>
     <string name="select_thumbnail">썸네일 이미지 선택</string>
 
-
+    <string name="track_size">트랙 %d개</string>
 
     <string name="error_message_connection">"error_message_connection"</string>
     <string name="error_message_ssl_hand_shake">"error_message_ssl_hand_shake"</string>
